requirements.txt
opencv-contrib-python
numpy
imutils
customtkinter
pillow